Simplified process

+++ DataHelpers/ForeignTableManager.cs
@@ -5,40 +5,49 @@ namespace MDR_Importer;
 
 public class ForeignTableManager
 {
+    private readonly Source _source;
     private readonly string _db_conn;
+    private readonly ILoggingHelper _logging_helper;
 
-    public ForeignTableManager(string db_conn)
+    public ForeignTableManager(Source source, ILoggingHelper logging_helper)
     {
-        _db_conn = db_conn;
+        _source = source;
+        _db_conn = _source.db_conn ?? "";
+        _logging_helper = logging_helper;
     }
 
-    public void EstablishMonForeignTables(string user_name, string password)
+    public void EstablishForeignMonTables(ICredentials creds)
     {
-        using var conn = new NpgsqlConnection(_db_conn);
-        string sql_string = @"CREATE EXTENSION IF NOT EXISTS postgres_fdw
+        if (creds.Username is not null && creds.Password is not null)
+        {
+            using var conn = new NpgsqlConnection(_db_conn);
+            string sql_string = @"CREATE EXTENSION IF NOT EXISTS postgres_fdw
                                  schema sd;";
-        conn.Execute(sql_string);
+            conn.Execute(sql_string);
 
-        sql_string = @"CREATE SERVER IF NOT EXISTS mon "
-                     + @" FOREIGN DATA WRAPPER postgres_fdw
+            sql_string = @"CREATE SERVER IF NOT EXISTS mon "
+                         + @" FOREIGN DATA WRAPPER postgres_fdw
                         OPTIONS (host 'localhost', dbname 'mon', port '5432');";
-        conn.Execute(sql_string);
+            conn.Execute(sql_string);
 
-        sql_string = @"CREATE USER MAPPING IF NOT EXISTS FOR CURRENT_USER
+            sql_string = @"CREATE USER MAPPING IF NOT EXISTS FOR CURRENT_USER
                  SERVER mon 
-                 OPTIONS (user '" + user_name + "', password '" + password + "');";
-        conn.Execute(sql_string);
+                 OPTIONS (user '" + creds.Username + "', password '" + creds.Password + "');";
+            conn.Execute(sql_string);
 
-        sql_string = @"DROP SCHEMA IF EXISTS mon_sf cascade;
+            sql_string = @"DROP SCHEMA IF EXISTS mon_sf cascade;
                  CREATE SCHEMA mon_sf; 
                  IMPORT FOREIGN SCHEMA sf
                  FROM SERVER mon 
                  INTO mon_sf;";
-        conn.Execute(sql_string);
+            conn.Execute(sql_string);
+            
+            _logging_helper.LogLine("Foreign (mon) tables established in database");
+        }
     }
 
 
-    public void DropMonForeignTables()
+    public void DropForeignMonTables()
     {
         using var conn = new NpgsqlConnection(_db_conn);
         string sql_string = @"DROP USER MAPPING IF EXISTS FOR CURRENT_USER
@@ -50,6 +59,77 @@ public class ForeignTableManager
 
         sql_string = @"DROP SCHEMA IF EXISTS mon_sf;";
         conn.Execute(sql_string);
+        
+        _logging_helper.LogLine("Foreign (mon) tables removed from database");    
+    }
+    
+    
+    public void UpdateStudiesImportedDateInMon(int importId, int? sourceId)
+    {
+        string top_string = @"Update mon_sf.source_data_studies src
+                      set last_import_id = " + importId + @", 
+                      last_imported = current_timestamp
+                      from 
+                         (select so.id, so.sd_sid 
+                         FROM sd.studies so ";
+        string base_string = @" ) s
+                          where s.sd_sid = src.sd_id and
+                          src.source_id = " + sourceId;
+
+        UpdateLastImportedDate("studies", top_string, base_string);
+    }
+
+    
+    public void UpdateObjectsImportedDateInMon(int importId, int? sourceId)
+    {
+        string top_string = @"UPDATE mon_sf.source_data_objects src
+                      set last_import_id = " + importId + @", 
+                      last_imported = current_timestamp
+                      from 
+                         (select so.id, so.sd_oid 
+                          FROM sd.data_objects ob ";
+        string base_string = @" ) s
+                          where s.sd_oid = src.sd_id and
+                          src.source_id = " + sourceId;
+
+        UpdateLastImportedDate("data_objects", top_string, base_string);
+    }
+
+
+    private void UpdateLastImportedDate(string tableName, string topSql, string baseSql)
+    {
+        try
+        {   
+            using NpgsqlConnection conn = new(_db_conn);
+            string feedbackA = $"Updating monitor import records, (mon.source_data_{tableName}), ";
+            string sqlString = $"select count(*) from sd.{tableName}";
+            int recCount  = conn.ExecuteScalar<int>(sqlString);
+            int recBatch = 100000;
+            if (recCount > recBatch)
+            {
+                for (int r = 1; r <= recCount; r += recBatch)
+                {
+                    sqlString = topSql + 
+                                 " and so.id >= " + r + " and so.id < " + (r + recBatch)
+                                 + baseSql;
+                    conn.Execute(sqlString);
+                    string feedback = feedbackA + r + " to ";
+                    feedback += (r + recBatch < recCount) ? (r + recBatch - 1).ToString() : recCount.ToString();
+                    _logging_helper.LogLine(feedback);
+                }
+            }
+            else
+            {
+                sqlString = topSql + baseSql;
+                conn.Execute(sqlString);
+                _logging_helper.LogLine(feedbackA + recCount + " records, as a single batch");
+            }
+        }
+        catch (Exception e)
+        {
+            string res = e.Message;
+            _logging_helper.LogError("In update last imported date (" + tableName + "): " + res);
+        }
     }
 }
